fix(Split In Batches Node): Roll back changes in v1 and create v2
